+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG
 
 * Support HTTPS as health check protocol (@timoreimann)
-* add controller to manage worker firewall for public access (@MorrisLaw)
+* Add controller to manage worker firewall for public access (@MorrisLaw)
+* Add prometheus metrics instrumentation to firewall controller (@MorrisLaw)
 
 ## v0.1.26 (beta) - June 16th 2020